Improvements to Cell View's cell list
Worldedit now emits signals when loading and unloading cells, and when changing the current cell.

The `rendered_landscape` class now uses landscapes' preexisting vertex normals.

Improvements to the Cell View window's cell list:

The columns are now resizable, and use flex sizing.

The checkbox to sort loaded cells above unloaded cells is now functional.

The current cell in the Render Window will be shown in bold.

Code cleanups and bugfixes, including to code which emits `dataChanged` when a row changes. (The region we marked as changed extended one column past the end, which confused QSortFilterProxyModel.)