Worldedit: Support for worldspace (un)loading
Worldedit can now maintain a 5x5 grid of loaded cells, when in a worldspace. Cells will dynamically unload and load as the camera position changes, to maintain a 5x5 grid centered on the camera's containing cell.

Worldedit now sets scene-global-state lighting/fog data only when the current area is set; not when loading or unloading individual cells. For worldspaces, the defaults used now reflect Skyrim's typical daytime sky.

Vulkan: a significant amount of debug logging related to mesh loading has been turned off. Mesh loading should no longer spam the logs, allowing logs related to other tasks (e.g. cell (un)loading) to actually be visible now.

---

Vulkan: the scene global state's fog-far-color is now used as the clear value when recording graphics command buffers. This allows for correct results when buffers are recorded; however, changes to the scene global state object won't result in real-time changes to the scene background color; you'll need to force a command buffer rebuild.

For now, this all has the intended effect, because we generally only change fog colors when changing cells, and changing cells also changes the objects loaded and so forces a command buffer rebuild. However, it won't work for things like "change the time of day" in the future.

We can improve this by going back to black as a clear color, and beginning our renders with a full-screen polygon that pulls its color from the scene global state via a shader; that will then allow for real-time updates to the fog color to also change the scene background color.

---

Vulkan: implemented a DISGUSTING HACK for landscapes to allow unloading and loading landscapes on the same frame without hitting the max landscape limit. This hack entails adding a `scene_frame_item_state::pending_reload` enum value, which is used only by landscapes, and allows a landscape that is pending-delete to be recycled within a single frame. After the `surface_renderer` carries out pending deletions, it will check for these pending reloads (for landscapes only) and if any are seen, it will force an update. This is possible for landscapes only because they currently have no per-frame data; all landscape information is coalesced into a single massive buffer.

The *right* implementation will be to decouple `rendered_landscape` indices from slots in the coalesced landscape buffer. We'll want to maintain a CPU-side grid of indices mapping the former to the latter, along with a count of how many landscapes are (or should be) active. We'll allow landscapes to become active above the max count; there simply won't be room for them in the index grid, so they'll never be coalesced. Whenever we're about to execute deletions of any pending landscapes, we'll also check the active count: if the number of active landscapes plus the number of pending-deletion landscapes exceeds the max grid count, then we know that some active landscapes were not coalesced due to lack of space; after pending deletions are carried out, we'll search for and coalesce any landscapes lacking a grid index, and give them a grid index at that time.

---

Vulkan: we now honor the "culled by application" flag on NiAVObject (listed in NifSkope as "Hidden"). This deals with WRDrawBridge01.nif, which for some godforsaken reason has an untextured mesh filled with corrupted polygons (roughly resembling the layout of the Whiterun bridge's gate, including a wooden ramp) which Bethesda just left in the file and flagged as hidden. Now that we honor the "hidden" flag, that mesh no longer displays in our Render Window, reducing the odds that I'll see it a few months from now and waste time trying to figure out if *I* did something wrong.

---

Added support for the following NIF block types seen around Whiterun's entrance: BSBlastNode, BSDamageStage, BSMultiBound, BSMultiBoundData, BSMultiBoundDataOBB, BSMultiBoundDataSphere, BSMultiBoundNode, BSRangeNode.

The "jump to grid position" widgets in the Cell View window are now functional.

Added natvis for the Vulkan `rendered_landscape` class, and for form stubs.

---

Helpers:

cobb::resizable_square_grid::for_each (implemented)
cobb::resizable_square_grid::shift_by (fixed)
natvis: cobb::resizable_square_grid
natvis: cobb::resizable_grid