Vulkan: (not fully tested) support for DDS colors in orders other than RGBA
Vulkan: natvis for DDS headers and pixel formats.

Vulkan: We now attempt to support uncompressed DDS files with color components not in RGBA order. Vulkan has no directly matching formats, but swizzling should allow it to work. To that end, image swizzle masks are now part of our Vulkan image metadata.

I've tested this in a debugger and it seems to work, but I need to make a specific test image and run a test "in the laboratory," on something I can actually see in action, to be sure.