Scene entities standardized; recycling implemented
Created the term "scene entity" to refer to anything that is used to draw a scene: `loaded_texture`, `rendered_bounds`, `rendered_landscape`, `rendered_light`, and `rendered_mesh`.

We now make extensive use of templates and metaprogramming in order to standardize how scene entities are created, updated, deleted, recycled, and synchronized with the GPU. This has substantially reduced code duplication and improved maintainability. As a result of these changes, I was also able to implement entity recycling (i.e. what's done with landscapes) for all entity types. Recycling is handled at the same time as pending deletions.

Made preparations for being able to change the size of the loaded outdoor cell grid at run-time. The `scene` object now has a config variable for it, though other pieces of code may still use the compile-time constant; the variable is also never changed.

Removed most of the initial test scene from `surface_renderer`, save for the test floor. I plan on eventually converting that into a configurable grid relying on a custom shader.

When `surface_renderer` attempts to load a texture by file path, it first checks if any of the already-loaded textures have that file path. The functionality for this search has been moved to a new member function on `scene`.

Fixed an oversight involving the refcounts on landscape textures: most of the textures used by a landscape would not have refcounts incremented when the landscape loaded, allowing those textures to potentially be deleted and replaced out from under the landscape as you flew around a worldspace.

Helpers: created `cobb::class_map`, which is similar to `cobb::class_array` except that it maps each class to a value of a given type; conceptually, it's like a hash map where the classes are keys.