Cleanup for scene entity overhaul; repo readme -> VS2022
Improved natvis for scene entities.

Removed unused files related to scene entities, including files for their older implementation.

Added a debug button to the Render Window that queues a __debugbreak() to occur at the start of the next frame draw, as an easy way to inspect the state of the `surface_renderer`.

Updated the repo readme and associated documentation to reflect the fact that I'm now using VS2022.