Landscapes now coalesced per-FiF; started work on configurable uGrids
Coalesced VIBs for landscapes are now stored per frame in flight, and are now updated using metaprogramming.

Wrote a prototypical, unused, untested class (`vulkanDK::scene_entities::coalesced_vib`) that could be used for coalescing non-fixed-length VIBs, with reuse, e.g. for `rendered_mesh` in the future.

Started work on making it possible to change the maximum landscape count in the renderer. To be blunt: it's going horribly. Resizing the GPU-side buffers requires a descriptor update, which the compiler is occasionally mishandling (for some godforsaken reason it's mangling the VkDescriptorBufferInfo::range field, but only some of the time). There's a much more damning oversight, however: landscape indices match 1:1 between the entity list and positions in the buffer, which means that if we *reduce* the grid size, then we'd have to forcibly unload and reload all landscapes, *from outside the renderer*, to get them into indices that "fit" within the shrunk buffer.

Resizable uGrids is still possible but the current implementation, with a flexible max count, is a wash. Instead, we should set some hard maximum landscape count (e.g. uGrids 25 would produce a hard maximum of 625) and then just let Worldedit load fewer than that at its discretion.