Preliminary implementation of landscape debug borders
The loaded-form class for Landscapes had helper functions for working with landscape vertex offsets. These have been made `constexpr`.

We now display landscape borders in a manner similar to the Creation Kit and Skyrim. This implementation is very preliminary. Borders are shown in the right place, but the two border colors aren't properly interpolated (the line is dashed rather than blended; I copied and pasted shader code that used a flat int to get this all working quickly). They also aren't pulled from the game INIs, and cannot be disabled yet. Finally, the changes made to our coalesced-VIB code to facilitate this feature are horribly disorganized and in dire need of a refactor.

To allow the borders to be hidden and the colors to be customized, we should have a global buffer (akin to `scene_global_state`; if desired, we could even put the fields on there in the first place) which stores the landscape border colors and a bool or flags-mask indicating whether borders should be shown. The vertex shader for borders should position all vertices off-screen if borders are meant to be hidden; this means that we won't need to re-record command buffers to toggle border visibility. Making the border colors into shader parameters will allow us to change them at run-time: we can read them from the game INIs, but we could also allow user customization.

As for the CPU side of things? We need to modify the coalesced VIB metadata, for scene entity types, to support specifying data for multiple "additional shared index buffers;" we may want to template it on a count in order to facilitate this. Additionally, we need to give it a member function that quickly computes size information, including the offset of the vertex buffer (accounting for necessary alignments); this size information should be similar to the private `_vib_size_info` struct used when updating VIB state, should replace that struct, and should also replace the two 99%-identical lambdas currently used for rendering landscapes (when binding the vertex and index buffers).

This commit also includes two enums in /dovah/data/ that were missing from previous commits; some of these enums are used for certain NIF blocks.