Improvements to rendered_landscape debugging
Fixed natvis for `rendered_landscape`: the blend grid is now shown correctly. Used this to confirm that we are properly copying the blend data over from the loaded form, and rearranging it properly for Vulkan.

Implemented natvis for scene entity handles and loaded LAND forms.

Fixed natvis for form stubs: the condition for when to show loaded form data was incorrect.

Implemented natvis for cobb::grid. Fixed natvis for cobb::resizable_square_grid.