Replaced test floor with test grid
Disabled the "test floor" mesh, and created a new debug grid. The grid is an XY plane which has adjustable line spacing and extents, displays as darker when viewed from below, and which is anchored to (0, 0, 0). It's rendered with alpha transparency and anti-aliases itself with the same; this can be disabled from C++.

We could in the future make the grid more configurable (e.g. with shader params) and use it for things like displaying the current snap grid, when snap-to-reference is enabled. For now, I just use it as a quick way to navigate the "void" when no cells are loaded.