Fixes for staging buffer soft cap; start of CPU-side benchmarking
Fixed errors in the implementation for staging buffer soft caps (including the implementation of deferring a scene entity's CPU-to-GPU resource transfer). The surface renderer didn't reset an entity's frame sync state (so frames-in-flight wouldn't trigger descriptor updates for textures); we also needed it to signal to frames-in-flight that a drawn entity's transfer (e.g. of its VIB) completed, so that command buffers could be rerecorded for meshes. With these changes, forcing the staging buffer soft cap to 1KB for debugging purposes allows us to see meshes and textures load in one by one when viewing a worldspace. The feature works properly, then.

Surface renderer now has debugging code for benchmarking the time it takes to scan over a NIF tree and generate `rendered_mesh`es and `loaded_texture`s from it.

Created `cobb::dummy`, `cobb::dummy_of`, and `cobb::dummy_type_if_false` as tools for creating no-op objects that just "eat" anything you pass into them. The principal use case is for real value or a dummy based on some debugging bool: using these dummies seems to prevent false-positive IntelliSense errors when invoking operator overloads from inside an if-constexpr conditioned to only run on the non-dummy. We use this in `surface_renderer` for some benchmarks that can be switched on or off at compile-time.