Moved NIF loading to renderer; prep for multi-threaded loading
Added an overload to `surface_renderer` which takes form model data and loads a NIF. This replaces a helper function used within Worldedit, and means that the loading of NIFs is now done entirely at the renderer's discretion; this will make multi-threading easier.

Removed duplicated Bethesda shader specularity copy code from `surface_renderer`.

Wrapped NIFs related to Worldedit and the Vulkan renderer in a `rendered_nif` class, which will be used to store state related to multi-threaded loading. Added a few checks to Worldedit to prevent it from trying to touch a NIF while it's being loaded off-thread.

Wrote some helper functions that we'll be using as part of the multi-threaded load process, though these functions are not themselves going to be invoked from off-thread.