FPS gains; rendered_nif threading redesign; bugfixes
Improvements to the rendered mesh shader -- correcting for an oversight and significantly improving frame rates.

When bringing NIF vertex data to `rendered_mesh` instances, we now ignore the "Vertex Alpha" shader flag if no NiAlphaProperty is set. This should be consistent with the game.

We now properly reset rendered meshes when they are recycled from a pending-deletion state.

The `rendered_nif` class is no longer a QObject subclass. We now use an atomic flags mask for its multi-threaded load state; this is an attempt to fix the previous, broken, attempt at thread-safety.