Worldedit syncs with SR; Cell View syncs with Worldedit
Improved how surface renderer stores its queues of assets to load in the background: a templated helper struct is used, making the code more organized.

Worldedit is now notified when the renderer finishes working with a batch of NIFs in the background. It uses this to update any bounds (for selected refs) that were generated without being able to rely on NIFs.

Fixed the "form type name to string" function: I had forgotten to list the "potion" form type. Added a static assertion to ensure form types are not missed when the code is maintained in the future.

Modified the Cell View window to synchronize its selection with Worldedit's selection as appropriate. There is an edge case here: if you filter the Cell View ref list, use Worldedit to select a ref that's filtered out, and then unfilter the list, Cell View will not have the ref selected.

Modified the Cell View window: the list of references now displays the base forms' types (e.g. STAT, ACTI) rather than the references' types (e.g. only REFR and ACHR, basically).