Cell View / Worldedit selection synchronization work
We should now be prepared for the case where if a ref is selected in Worldedit and filtered out from display in Cell View, and the ref is edited such that it's no longer filtered out, then we select it in Cell View. We don't currently have the means to test this, however.

The ref list in Cell View now uses a DKHeaderView, so columns size better now.