Windows undef header overhauled; Worldedit reorganized; edit gizmo
I had a header which #undefs bad macros in the Windows headers. I decided to overhaul and replace it: I now have a header which includes <Windows.h> and then systematically replaces all of the FunctionA/FunctionW macros with `auto&` references. For <Windows.h> in its default configuration, this should solve the macro pollution issue completely. For alternate configurations, I'm less confident.

Reorganized Worldedit to match the Worldinput and XInput subsystems.

Added an edit gizmo to the `surface_renderer`. The mesh is generated programmatically and can be configured through code; a struct similar to `scene_global_state` can be used to update aspects of the gizmo. Currently, only a "translate" model is defined (no "rotate" or "scale" models yet), and the gizmo is not interactive. Whether and where the gizmo renders can be controlled by outside code e.g. Worldedit, though currently, only a test button on the Render Window is wired to it.

In order to support rendering the gizmo, we needed a new render pass (so that gizmos could render with a fresh depth buffer), a new VIB, and a new descriptor group (set and set layout) for the `scene_gizmo_state` buffer.

Additionally, I improved how the "render passes by name" list is synched into the `abstract_renderer::render_passes` list, though at some point, I should really see if there's a better way to handle that -- some way of getting the benefits that would be easily achievable through CRTP, without the downsides i.e. not being able to hide/contain includes in a CPP file.