Edit gizmo models can now swap without rebuilding command buffers
We now coalesce all edit gizmo models into a single VIB, and use the vectors' W-component to indicate which model they belong to. Code in the vertex shader will position a vertex off-screen if it belongs to a model other than the one we currently want to render. In this way, we can change the edit gizmo's current model without having to ever rebuild command buffers.

Fixed a bug with how the "translate" gizmo's mesh was generated: we were generating way too many indices, effectively rendering tons of triangles consisting of vertices 0, 0, and 0.