Form editing dialog overhaul; gui_adjust
Refactored the base classes for form editing dialogs.

Added a `gui_adjust` subsystem. This subsystem will allow adjustments that a user makes to list view column widths in one dialog to persist through the same list view in all dialogs of that type, for the rest of the session. Adjustments become persistent when the dialog in which they were made is closed.

Modified Worldedit and associated classes to use surface renderer raycasts for mouse testing.