Work on the Worldinput rewrite
Refactored the bind tree and its nodes.

Sketched out the bind tree per-frame update algorithm. Several changes are still needed and are marked with static assertions.

Created an `editor_mode` enum in Worldedit, which is referenced from Worldinput. Currently, Worldedit doesn't use this enum, but we'll need to add a getter soon in order for editor mode nodes in Worldinput to work.

Separated-and-ordered input sequence groups now handle the key sequence expire time.