Work on the Worldinput rewrite
All input device handlers now inherit from an abstract base class, with virtual member functions for getting device button state and for consuming buttons.

The input sequence update algorithm now has access to the current input device, and uses this access to query button state and consume buttons as necessary.

The bind tree per-frame update algorithm now receives the current timestamp, and passes this to the input sequence update algorithm. Implemented recursion for the tree traversal step of the algorithm. Implemented matching for editor-mode nodes. Renamed the `input_node` variable to `child_inode` to make its relationship to `child` (and to the algorithm as written in prose) clearer. Implemented the assertion that all groups in a released input sequence be inactive at the time of releasing.

Worldedit now has a state value indicating the editor mode, and a getter for this state value. Currently, there is no way to switch to other editor modes.