Worldinput rewrite: absolute input sequences avoid redundant ISGs
When creating an absolute input sequence given a "parent" sequence and a "child" sequence: if the parent sequence ends in an ordered group, we can now merge the child sequence directly into that group, avoiding the recreation of redundant nested groups (e.g. an concurrent-ordered group as the child of another concurrent-ordered group).