Worldinput rewrite: fixed press-delays-hold; added singleton
Fixed the press-delays-hold conflict resolution rules. Previously, we specified hat a (Long) Press input node can only conflict with a Hold input node if they have the same absolute terminal inputs. However, that meant that cases like `Long Press [A + B]` versus `Hold [A + C]` wouldn't be considered a conflict. The spec and code have been rewritten so that we detect a conflict if the absolute terminal inputs overlap.

Created the `worldinput2` singleton. Large portions of code are borrowed from the original `worldinput` singleton, but we've trimmed it down a lot and also removed some functionality that is unused even in the original.

Created a placeholder type for `combined_tool_results`.