Worldinput rewrite: ISG last advancement time moved to ISes
Moved the "last advancement time" state value from input sequence groups to whole input sequences, and modified the update algorithms for both to allow that to work. ISGs now signal when they have "advanced" in any capacity, and this is used to update the last advancement time on the input sequence as a whole. In turn, that timestamp is passed to ISGs as a parameter, and separate-and-ordered groups use it to timeout input sequence entry if the user takes too long to move from one child ISG (e.g. one key) to the next.