Worldinput rewrite: separate-and-ordered interruption check
Designed and implemented an algorithm by which we check for the interruption of an in-progress separate-and-ordered ISG by an unrelated key. That is: if you're in the middle of entering, say, the Konami code, and you accidentally press Left Bumper, then that should interrupt the key sequence such that it fails to match and you have to re-enter it from the start.

I have no way of knowing if this will even work until I begin specifically testing it, but I'm semi-confident in it.