Worldinput rewrite: recursive node disqualification
When traversing the bind tree, if we disqualify a hold node, then we now also recursively disqualify its descendants.

In the future, I may just have bind trees be "baked" into flat lists; it'll greatly simplify the process of iterating over and evaluating the tree, and it may improve memory locality as well. (This is relevant here because it means that if a modifier node loses a conflict with another node, all of its children should theoretically have and lose the same conflict.) For now, though, the trees are actually *stored* as trees and must be iterated accordingly.