Worldinput rewrite: nested first-child edge-case solved
The input sequence [B + [A + B]] should be impossible. Previously, however, you could trigger it by inputting [A + B]. This is because the inner sequence was not cognizant of the down timestamp of its previous sibling. The ISG update algorithm has been revised to remedy this.

In order to fix additional edge-cases stemming from the above change, device button state now includes the down timestamp for a key released on the current frame (we've decoupled the fact of a key currently being down from us tracking when it *went* down); and when an ISG update call returns the "released" frame status, it now also returns the timestamp at which the released ISG originally went down.