Worldinput rewrite: s/o group first child honors previous-sibling timestamp
When a separate and ordered ISG is processing its first child, it now requires that that child have gone down at or after the previous-sibling timestamp. This means that the sequence [X + <A + S + J>] can no longer be completed with the following inputs: A down; X down; A up; S down; S up; J down; J up.