Worldinput rewrite: start of tree-flattening approach
The tree structure is useful for defining control schemes, but isn't needed to actually handle inputs. 90% of our spec'd behaviors are irrelevant to whether we use a tree or a flat list, so let's make the spec and the overall design simpler and use a flat list for input processing.

(Plus, it'll play well with some optimizations I want to make.)

Started work on writing the code for flat "bind lists." This commit includes the basic bind list data structures, all conflict resolution algorithms, the bind list update algorithm, and the algorithm for flattening a bind tree into a bind list.