Worldinput rewrite: work on directional inputs
Input sequence stringification should now support directional input constraints, though this is not yet tested.

For the same-frame concurrent binds conflict resolution, when two binds have unequal specificity and the same directional input, they are considered to be in conflict. We only care about same-frame conflicts for DIs, and changing the modifier rule (whether to make an exception when DIs are unequal, or to only apply it when they're equal) feels like it'll lead to unintuitive behavior.