Worldinput rewrite: Range constraints finished
Renamed "directional constraints" to "range constraints," given the nomenclature wherein scalar and vector input controls are collectively "range input controls."

Created the term "delta control" for a control that has no consistent point of reference. The main example of a delta control is mouse movement: if all we care about is that the mouse is moving in a leftward direction, then "leftward" is defined in terms of a variable point of reference (the mouse's prior on-screen position, which can be anywhere).

It is possible for us to poll the state of a delta range control faster than the OS can (or is willing to) update it. This means that even continuous mouse movement at average speeds can still lead to several ticks on which the mouse appears to have remained still. We work around this as follows: a delta range control is considered "stale" for a brief period after it stops moving. Stale range controls satisfy range constraints even though the controls are zeroed; however, a bind with a range constraint will not have its bound tool invoked on any frame on which the associated range control is stale (to prevent no-op invocations with zeroed parameters).

With the above changes, range controls appear to work exactly as desired in tests.