Worldinput rewrite: Minor organizational change
The bind list is no longer directly responsible for testing range constraints; rather, it uses an `input_sequence::range_constraint` member function to do it.

The reason this isn't integrated into `input_sequence::update` is because we want a Hold bind to clear its Press-blocked-Hold flag when the buttons are released, but not when the range constraint ceases to be met. Range constraints are not intended to influence the handling of button press type conflicts, nor can they do so reliably: a very slow mouse movement, for example, may be a contiguous physical movement, but because mouse coordinates are measured in integer pixels, it will register as a series of starts and stops depending on the timing at which the cursor crosses pixel boundaries; and so such a movement would improperly clear the Press-blocked-Hold flag if range constraint failures allowed such flags to be cleared.