Worldinput rewrite: started work on raycast requirements for input sequences
Added `raycast_requirement` and `raycast_result` structs to Worldinput. They are not currently used in any processing.

Added fields to `input_sequence` which can be used to define a raycast requirement and associate it with exactly one group in the sequence (which our design currently requires to be a button).

Added fields to the Worldinput2 core to track per-frame raycast results, for both the mouse cursor and the on-screen reticle used by gamepads (though DovahKit does not currently draw a reticle).

Added a field to `tool_invocation_cause` for raycast results.

Added a compile-time flag to Worldedit tools that signals them as reacting to the results of a raycast. Added compile-time checks to tools to ensure consistent configuration for tools that have this flag set. Set the flag and all configuration options it requires, for two existing tools that will be rebuilt to use the raycast result system.