Worldinput rewrite
The `input_sequence::group::update` function now receives its arguments wrapped in a params struct. We're needing to pass more and more state to it, but don't want it to have full access to the containing input sequence or its internal state, so this should help make that more manageable.