Worldedit/Worldinput work: edit gizmo behaviors and tools
Worldedit now manages edit gizmo visibility and position consistently with the Creation Kit. We blindly update it every frame, which probably isn't the most efficient way to do things, but it works for now.

The surface renderer now offers a transform setter for the edit gizmo, which properly updates internal change-flags to ensure CPU-to-GPU synchronization on demand.

Created a Worldedit tool for toggling the edit gizmo's visibility, mode, and reference frame. The Debug WASD control scheme has had binds added for this tool, though unfortunately, those actually conflict with WASD and will need to be moved to a new "CK-like" control scheme (which we want for release anyway) in order to allow for comfortable testing.

Fixed an error in the surface renderer's background loader, which could cause it to accidentally underflow a counter for entities pending GPU upload.

Edit gizmos now use their own command buffer instead of sharing a command buffer with rendered bounds. This fixes a conflict wherein rendering any of the latter would prevent the gizmo from displaying, since each buffer resets, begins, and ends, and the two are handled in different functions which both perform those steps.