Bitstream redesign in progress
The old bitstream system wasn't easy to test, was impossible to debug, isn't working, is error-prone, and has no built-in facilities for data versioning or run-time error handling. It wasn't terrible for a first draft, but it's in dire need of a redesign; thus, this commit.

The new bitstreams live in `cobb::bitstreams::*` and are designed to store a data header, throw exceptions on errors, and use more symmetric interfaces. Compile-time tests exist for basic values (including various bit offsets), structs, and containers. Unused functions have been pruned.