Bitstream work
Bitstream objects are capable of tracking what structs they're inside of -- essentially, an approximate mirror of the current call stack. Macros and metaprogramming are used to get structs' fully-qualified names. This information can be queried during serialization, and is also stored in any thrown exceptions.

Added a specializable trait struct that can be used to override the bitcount of any enum, and to supply a list of enum values that are considered valid (for enum validation during reads).

Created a `static_warning` macro that can be used to emit compile-time warnings based on language-level (rather than pragma/preprocessor) expressions. Unfortunately, it suffers from some issues related to the limitations of C/C++'s preprocessor, and while workarounds exist, none of them appear to work in IntelliSense.