Bitstream headers now version the stream internals and data separately
Bitstream headers now store a bitstream version and a data version. The former is me attempting to future-proof the bitstream machinery itself, in case there are features we add or fixes we make. The latter would be used by whatever's actually constructing and using the bitstreams.

Added better documentation to `cobb::bitstreams::enum_serialization_options`.