UI work
Implemented a `sibling` override on `DKWorldinputInputSequenceModel`; this is needed for `QModelIndex::siblingAt` to work properly.

The input sequence bind editor now displays an icon on the raycast-associated button, and offers a context menu.

Unfortunately, something is wrong with row selection for the input sequence bind editor's treeview. Qt's different systems are acting inconsistently with one another, which makes me suspect that it's an issue with the Qt model boilerplate -- despite my having copied most of that boilerplate from previous models that actually work. Qt's boilerplate is near-impossible to debug (not least because Qt's `testlib` breaks Qt VS Tools), so I've started work on a CRTP-based model class that will automate most of the boilerplate for me -- and separate it away from the actual data-specific logic.