Compiler update fixups; bind editor can now set tool options
Made several changes needed to allow the project to build after a VS/MSVC update:

Disabled compiling of the STL modules. MSVC does not properly handle build paths with them; the trick one would use to allow relative object file paths (so that the linker doesn't choke when two files in different places have the same name) produces broken paths for the modules.

Even if you `using enum` in a scope, the names of that enum's members can conflict with the names of any namespaces that are in-scope. The precise reason for this, and the precise operational definition of "in-scope," are not clear, nor is it clear to me whether this is a language limitation that MSVC did not previously enforce or a new bug in MSVC. This required changes to the serialization configurations for some Worldedit enums.

Fixed various places where MSVC now needs me to use the `typename` keyword on a dependent name. Notably, MSVC emits C2275 with message "
expected an expression instead of a type," which is literally the wrongest possible message to display in these situations. Like, impressively wrong.

The operator== overload for scene entity handles in my renderer started breaking inexplicably, and MSVC failed to provide any useful diagnostic information (though it did provide a list of literally every single non-member operator== overload across the entire project, including in Qt, so that's... cool, I guess). I fixed this by switching the operator to use deducing this rather than CRTP. I have no idea why this even worked.

Removed a now-unnecessary comparison operator overload from `DKKeyPickerWidget::_key`, used to allow comparisons to the superclass without casting.

---

Fixed the enum combobox helper: the pair members were backwards in some code, among other mistakes.

Removed default padding from several wrapper widgets, and made other layout fixes, for Worldedit tool options.

---

The Worldinput bind editor now allows configuration of tool options. This entailed the creation of a dispatch table for creating a tool's options widget, and for getting data into and out of the widget.