Work on a generic node class
I want to redesign how control schemes are stored, as the current design is causing some logistical and other issues in implementing the UI model. Basically, there's not really a "clean" way to implement the model, because for control schemes we treat nodes and their data as the same thing; nodes are polymorphic, and the data inside of a node isn't separable from the entire node (i.e. a heap-allocated thing with hierarchy).

As such, I'm trying to write a generic, metaprogrammed node type that acts as a container for data types defined elsewhere.