Progress on control scheme editor dialog
You can now view and edit the contents of an "action" node in a control scheme. The buttons to move and delete nodes also work.

Next steps are: dialogs for editing condition and modifier nodes; and implementing the ability to create new nodes inside (or after, if it's a leaf) the currently selected node.