Complete rewrite of (at the time unused) INI handling; start of options INI work
This commit is too large.

---

Completely rewrote all of my old INI-handling code from ReachVariantTool (and probably my old Skyrim LE DLLs), with the intention of using it for an INI file in DovahKit.

As part of that rewrite:

 - Used JS to generate code to check if any Unicode glyph is "other"-type (i.e. !QChar::isPrint).

 - Wrote helper functions to parse integer and floating-point literals.

 - Wrote helper functions for variadic types and similar, which ended up being unnecessary but which I'm going to keep in my toolbox. Ditto for a few index sequence and concept helpers.

 - Wrote basic functions for iterating over a std::string as UTF-8, by code point, and for appending a UTF-8 code point to a UTF-8 string.

 - Wrote functions for (un)escaping a string with backslashes, which assume UTF-8 and so rely on the above.

 - Created a throwaway testing project, currently used to probe the behavior of GetPrivateProfileString and GetPrivateProfileInt.

The end result is that the code I use to handle INI settings has grown considerably, but that code is now decoupled from Qt and runnable at compile-time. The code for parsing INI files can now be tested with static assertions, as can some of the code for handling INI settings in memory.

---

Created an INI file definition for DovahKit. It is currently unused.